requirements.txt
playwright
pytest
pytest-bdd
allure-pytest
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0 